requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.2
drf-yasg==1.21.7